Dont mess with size if its fine @RAM gear

+++ system/engine/gears/006pqr.sh
@@ -3,19 +3,20 @@
 
 B=/system/engine/bin/busybox
 
-if [ ! -h /data/local/tmp/adreno_config.txt ]; then
+if [ -e /system/lib/egl/libGLESv2_adreno200.so ]; then
+ $B sleep 1
+ if [ ! -h /data/local/tmp/adreno_config.txt ]; then
  $B chmod 777 /system/engine/assets/adreno_config.txt
  $B ln -s /system/engine/assets/adreno_config.txt /data/local/tmp/adreno_config.txt
+ fi;
+ $B mount -t debugfs debugfs /sys/kernel/debug
+ $B chmod 666 /dev/kgsl-3d0
+ $B chmod 666 /dev/msm_aac_in
+ $B chmod 666 /dev/msm_amr_in
+ $B chmod 666 /dev/genlock
+ $B chmod 777 /dev/graphics/fb0
+ $B echo 0 > /sys/kernel/debug/msm_fb/0/vsync_enable
+ $B echo 16 > /sys/kernel/debug/msm_fb/mdp/mdp_usec_diff_treshold
+ $B echo 30 > /sys/kernel/debug/msm_fb/mdp/vs_rdcnt_slow
 fi;
-
-$B mount -t debugfs debugfs /sys/kernel/debug
-$B chmod 666 /dev/kgsl-3d0
-$B chmod 666 /dev/msm_aac_in
-$B chmod 666 /dev/msm_amr_in
-$B chmod 666 /dev/genlock
-$B chmod 777 /dev/graphics/fb0
-$B echo 0 > /sys/kernel/debug/msm_fb/0/vsync_enable
-$B echo 16 > /sys/kernel/debug/msm_fb/mdp/mdp_usec_diff_treshold
-$B echo 30 > /sys/kernel/debug/msm_fb/mdp/vs_rdcnt_slow
-
